// src/types/hero.ts
export type Hero = {
  id: string;

  pre_heading: string;

  title_line_1: string;

  description: string;

  hero_image: string | null;

  stat_1_number: string;
  stat_1_label: string;

  stat_2_number: string;
  stat_2_label: string;

  stat_3_number: string;
  stat_3_label: string;

  stat_4_number: string;
  stat_4_label: string;
};